PHON-33: Add ability to upload video

+++ resources/views/mediafile/create.blade.php
@@ -21,21 +21,47 @@
                         <input type="submit">
                     </form>
 
+                    <!-- TODO factor out shared code into template -->
                     @foreach($mediafiles as $mediafile)
-                        <div class="row" style="margin:5px">
-                            <div class="col-md-3 col-md-offset-1">
-                                <img src="{{ URL::asset('/') . '/public/mediafiles/' . $mediafile->filename }}" style="max-width:100px; max-height:100px"/>
-                                {!! Form::open([
-                                    'method' => 'DELETE',
-                                    'route' => ['mediafile.destroy', $mediafile->id],
-                                    'class' => 'btn-group'
-                                ]) !!}
+                        @if ( $mediafile->mime == 'video/quicktime')
+                            <div class="row" style="margin:5px">
+                                <div class="col-md-7 col-md-offset-1">
+
+                                    <video controls="controls" style="max-width:400px" width="100%" height="auto" name="Video Name" src="{{ URL::asset('/') . '/public/mediafiles/' . $mediafile->filename }}"></video>
+
+                                    {!! Form::open([
+                                        'method' => 'DELETE',
+                                        'route' => ['mediafile.destroy', $mediafile->id],
+                                        'class' => 'btn-group'
+                                    ]) !!}
+                                        <button title="Delete Media" id="delete{{ $mediafile->id }}" type="submit" class="btn btn-danger" style="display:inline">
+                                            <span class="glyphicon glyphicon-remove" style="margin:-5px"></span>
+                                        </button>
+                                    {!! Form::close() !!}
+                                </div>
+                            </div>
+                        @endif
+                    @endforeach
+
+                    @foreach($mediafiles as $mediafile)
+                        @if ( $mediafile->mime != 'video/quicktime')
+                            <div class="row" style="margin:5px">
+                                <div class="col-md-7 col-md-offset-1">
+
+                                    <img src="{{ URL::asset('/') . '/public/mediafiles/' . $mediafile->filename }}" style="max-width:100px; max-height:100px"/>
+
+                                    {!! Form::open([
+                                        'method' => 'DELETE',
+                                        'route' => ['mediafile.destroy', $mediafile->id],
+                                        'class' => 'btn-group'
+                                    ]) !!}
                                     <button title="Delete Media" id="delete{{ $mediafile->id }}" type="submit" class="btn btn-danger" style="display:inline">
                                         <span class="glyphicon glyphicon-remove" style="margin:-5px"></span>
                                     </button>
-                                {!! Form::close() !!}
+                                    {!! Form::close() !!}
+                                </div>
                             </div>
-                        </div>
+                        @endif
                     @endforeach
                 </div>
